[FEATURE REQ] Relocate auth-functions to serverly.user.auth Closes #33 Also organized session-related functions to serverly.user.sessions
